barakah_app/backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-admin-interface==0.29.4
django-colorfield==0.12.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
